refactor: move dummy graphql results to separate file
