+++ backend/graph/schema.resolvers.go
@@ -50,34 +50,3 @@ func (r *Resolver) Query() QueryResolver { return &queryResolver{r} }
 
 type processResolver struct{ *Resolver }
 type queryResolver struct{ *Resolver }
-
-// !!! WARNING !!!
-// The code below was going to be deleted when updating resolvers. It has been copied here so you have
-// one last chance to move it out of harms way if you want. There are two reasons this happens:
-//   - When renaming or deleting a resolver the old code will be put in here. You can safely delete
-//     it when you're done.
-//   - You have helper methods in this file. Move them out to keep these resolver files clean.
-var dummyProcesses = []*model.Process{
-	{
-		ProcessKey:           1,
-		ProcessID:            123,
-		Version:              1,
-		BpmnLiveStatus:       "alive",
-		BpmnResource:         "a8s7dft87agvsdf=",
-		DeploymentTime:       "2023-01-01T00:00:00Z",
-		Instances:            []*model.Instance{},
-		MessageSubscriptions: []*model.MessageSubscription{},
-		Timers:               []*model.Timer{},
-	},
-	{
-		ProcessKey:           2,
-		ProcessID:            222,
-		Version:              3,
-		BpmnLiveStatus:       "dead",
-		BpmnResource:         "a7s6dtf7aiw3=",
-		DeploymentTime:       "2023-01-01T00:00:00Z",
-		Instances:            []*model.Instance{},
-		MessageSubscriptions: []*model.MessageSubscription{},
-		Timers:               []*model.Timer{},
-	},
-}
